change defaultConfig.ts
