+++ src/components/layout/defaultConfig.ts
@@ -1,13 +1,7 @@
-import { StackLayoutOptions } from "./StackLayoutConfig";
+import { createDefaultConfig, StackLayoutOptions } from "./StackLayoutConfig";
 
 const defaultConfig = (config?: StackLayoutOptions )=>{
-    const options = {
-        eventHandler: config?.eventHandler?? "DefaultHandler",
-        layoutStore: config?.layoutStore?? "DefaultLayoutStore",
-        dataStore:  config?.dataStore?? "DefaultDataStore",
-      } as StackLayoutOptions
-      
-    return { options }
+  return createDefaultConfig(config);
 }
 
 export  { defaultConfig }
